pyproject.toml
[tool.poetry]
name = "rightshero"
version = "0.1.0"
description = ""
authors = ["ahmedgamal"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.3"
djangorestframework = "^3.14.0"
uvicorn = "^0.27.1"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
